+++ engine/include/cubos/engine/render/camera/camera.hpp
@@ -1,0 +1,28 @@
+/// @file
+/// @brief Component @ref cubos::engine::Camera
+/// @ingroup render-camera-plugin
+
+#pragma once
+
+#include <glm/mat4x4.hpp>
+
+#include <cubos/core/reflection/reflect.hpp>
+
+#include <cubos/engine/api.hpp>
+
+namespace cubos::engine
+{
+    /// @brief Generic component to hold the projection matrix of a specific camera (either perspective or orthogonal).
+    /// @note Added automatically once a specific camera is added (e.g @ref cubos::engine::PerspectiveCamera).
+    /// @ingroup render-camera-plugin
+    struct CUBOS_ENGINE_API Camera
+    {
+        CUBOS_REFLECT;
+
+        /// @brief Whether the camera is drawing to a target.
+        bool active{true};
+
+        /// @brief Projection matrix of the camera.
+        glm::mat4 projection{};
+    };
+} // namespace cubos::engine